Add option to write paths to FIFO instead of open
The -o|--output-paths option specifies the path of a named pipe to open
or create. When this option is specified, the open bind (RVK_OPEN) will,
instead of opening the selected file with $OPEN, write the path of the
file to the named pipe, terminated by a '\0'. If there are any files
marked, their paths will be printed instead.

This option could be used together with an editor that could read paths
from the named pipe and open them for editing.